refactor: move check if symbol is from default lib or not to new fn
